--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/GameList.java
@@ -1,7 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-import java.util.ArrayList;
-
-public class GameList extends RestCall {
-  public ArrayList<GameData> games;
-}
