--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/GameRequest.java
@@ -1,7 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-public class GameRequest extends RestCall {
-  public String fromCell;
-  public String toCell;
-  public String username;
-}
