--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/InviteData.java
@@ -1,4 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-public class InviteData extends DataType {
-}
