--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/InviteList.java
@@ -1,7 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-import java.util.ArrayList;
-
-public class InviteList extends RestCall {
-  public ArrayList<InviteData> invites;
-}
