--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/QueryRequest.java
@@ -1,6 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-public class QueryRequest extends RestCall {
-  public String username;
-  public String gameId;
-}
