--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/UserData.java
@@ -1,4 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-public class UserData extends DataType {
-}
