--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/UserList.java
@@ -1,7 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-import java.util.ArrayList;
-
-public class UserList extends RestCall {
-  public ArrayList<UserData> users;
-}
