--- server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/UserRequest.java
@@ -1,7 +1,0 @@
-package edu.colostate.cs.cs414.f18.the_other_alex.server;
-
-public class UserRequest extends RestCall {
-  public String username;
-  public String email;
-  public String password;
-}
