+++ server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/reqjson/QueryRequest.java
@@ -1,0 +1,8 @@
+package edu.colostate.cs.cs414.f18.the_other_alex.server.reqjson;
+
+import edu.colostate.cs.cs414.f18.the_other_alex.server.RestCall;
+
+public class QueryRequest extends RestCall {
+  public String username;
+  public String gameId;
+}
