+++ server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/reqjson/UserRequest.java
@@ -1,0 +1,9 @@
+package edu.colostate.cs.cs414.f18.the_other_alex.server.reqjson;
+
+import edu.colostate.cs.cs414.f18.the_other_alex.server.RestCall;
+
+public class UserRequest extends RestCall {
+  public String username;
+  public String email;
+  public String password;
+}
