+++ server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/resjson/UserData.java
@@ -1,0 +1,6 @@
+package edu.colostate.cs.cs414.f18.the_other_alex.server.resjson;
+
+import edu.colostate.cs.cs414.f18.the_other_alex.server.DataType;
+
+public class UserData extends DataType {
+}
