+++ server/src/main/java/edu/colostate/cs/cs414/f18/the_other_alex/server/resjson/UserList.java
@@ -1,0 +1,9 @@
+package edu.colostate.cs.cs414.f18.the_other_alex.server.resjson;
+
+import edu.colostate.cs.cs414.f18.the_other_alex.server.RestCall;
+
+import java.util.ArrayList;
+
+public class UserList extends RestCall {
+  public ArrayList<UserData> users;
+}
